use derive for HasAttachment
